Feedback push!
Fixed special case for method "numberOfMatches()" where it would return the wrong amount if "firstIndexOf" and "lastIndexOf" returned -1.

Fixed the complexity on "firstIndexOf" and "lastIndexOf" where it now does not does a linear search when the key is found.

In the constructor of "Autocomplete" we now sort the array there by lex instead of sorting the array every time in the methods to that class.

Lastly "answers.txt" is updated for these changes.